transactions/requirements.txt
python-dotenv
psycopg2
sqlmodel
PyJwt
fastapi[standard]
bcrypt